svg file: temporarily fix the wrong headers
An SVN file is a valid XML file, but it is mostly handled by a graphic
program.
It seems the metadata should be added though the editor using some
standard SVG fields, but it is not fully clear (at least to me)
which ones are the correct ones, see also:

https://github.com/fsfe/reuse-tool/issues/180

In any case, the SPDX header before the xml line makes this
an invalid SVG file (even though some relaxed parser may still
accept it), so remove the SPDX comment for now.

CCMAIL: jbb@kaidan.im