Handle thumbnails that are not stored in the guessable location
Currently we just guess the url, which almost always works, but for some
videos (maybe very old ones) it doesn't work.
So for those, we now just invoke yt-dlp to extract the correct url. As
long as this affects only very few videos, it's still fast enough.

AudioTube will print a warning on the console if that happens, to make
it easier to debug when its slow.