update metadata to be compliant with flathub's quality guidelines
See https://phabricator.kde.org/T17275 and https://carlschwan.eu/appdata/org.kde.audiotube.json. This MR adds light and dark brand colors and adapts the summary and description.
Here is a preview of the proposed branding colors (made with https://docs.flathub.org/banner-preview/):
![branding-color-preview.png](/uploads/fdca549d60537a665f4b3c5560f4e847/branding-color-preview.png)

The other issue was that the summary was too long, so this MR removes the mention of convergence and move it to the description instead. I also updated the french and german translations.